+++ megalinter/MegaLinter.py
@@ -163,7 +163,7 @@ class Megalinter:
         if config.get("PARALLEL", "true") == "true" and len(active_linters) > 1:
             self.process_linters_parallel(active_linters, linters_do_fixes)
         else:
-            self.process_linters_serial(active_linters, linters_do_fixes)
+            self.process_linters_serial(active_linters)
 
         # Update main MegaLinter status according to results of linters run
         for linter in self.linters:
@@ -208,7 +208,7 @@ class Megalinter:
         self.check_results()
 
     # noinspection PyMethodMayBeStatic
-    def process_linters_serial(self, active_linters, _linters_do_fixes):
+    def process_linters_serial(self, active_linters):
         for linter in active_linters:
             linter.run()
 
@@ -217,14 +217,26 @@ class Megalinter:
         if linters_do_fixes is True:
             # Group linters by descriptor, to avoid different linters to update files at the same time
             linters_by_descriptor = {}
+            linter_groups_without_fixes = []
             for linter in active_linters:
-                descriptor_active_linters = linters_by_descriptor.get(
-                    linter.descriptor_id, []
-                )
-                descriptor_active_linters += [linter]
-                linters_by_descriptor[linter.descriptor_id] = descriptor_active_linters
+                if linter.apply_fixes is True:
+                    # If the linter can update sources, it must be runned in the same group than
+                    # other linters that can update the same sources
+                    descriptor_active_linters = linters_by_descriptor.get(
+                        linter.descriptor_id, []
+                    )
+                    descriptor_active_linters += [linter]
+                    linters_by_descriptor[
+                        linter.descriptor_id
+                    ] = descriptor_active_linters
+                else:
+                    # If the linter can not updates sources, no need to run it in the same group
+                    linter_groups_without_fixes += [[linter]]
+            # Add groups of linters that can update sources
             for _descriptor_id, linters in linters_by_descriptor.items():
                 linter_groups += [linters]
+            # Add "groups" of 1 linter than can not update sources
+            linter_groups += linter_groups_without_fixes
         else:
             # If no fixes are applied, we don't care to run same languages linters at the same time
             for linter in active_linters: